oracle/requirements.txt
requests>=2.31.0
web3>=6.0.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
odds-api-io>=0.1.0
